requirements.txt

pandas
numpy
scikit-learn
xgboost
scipy
networkx
spacy
bs4
requests

langchain
langgraph
langchain_community
langchain_openai
faiss-cpu
sentence-transformers
yfinance
fastapi
uvicorn
pydantic
requests
streamlit
Utilities
python-dotenv